version 0.1.0

+++ app/main.py
@@ -1,24 +1,26 @@
 import argparse
-# import sys
+import os
 
+from block import BlockingManager
+from commands import BlockSiteCommand
+from commands import ListBlockedSitesCommand
+from commands import UnblockSiteCommand
 from utils import copy_file, get_hosts_path
-# from block import BlockingManager
-# from block import ListBlockedSitesCommand
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
-        description="Block, unblock, list, or restore websites via the hosts file."
+        description="Block, unblock or list websites via the hosts file."
     )
     parser.add_argument(
         "action",
-        choices=["block", "restore", "unblock", "list"],
-        help="Action to perform.",
+        choices=["block", "unblock", "list"],
+        help="Actions to perform.",
     )
     parser.add_argument(
         "target",
         nargs="?",
-        help="The website to block/unblock (e.g., facebook.com) or '-all'.",
+        help="The website to block/unblock (e.g., facebook.com) or '--all'.",
     )
     parser.add_argument(
         "-m", "--minutes", type=int, help="Duration in minutes for temporary blocking."
@@ -32,54 +34,41 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
     hosts_file = get_hosts_path()
-    copy_file(hosts_file, r"../data/original_hosts")
-    # blocking_manager = BlockingManager(hosts_file)
+    command = None
 
-    # try:
-    #     with open(hosts_file, 'r') as f:
-    #         original_hosts_content = f.read()
-    # except IOError as e:
-    #     print(f"Error reading the hosts file: {e}")
-    #     sys.exit(1)
+    if not os.path.exists(r"../data/original_hosts"):
+        try:
+            copy_file(hosts_file, r"../data/original_hosts")
+        except Exception as e:
+            print(f"An error occured during copy: {e}")
+    blocking_manager = BlockingManager(hosts_file)
 
-    # command = None
-
-    if args.action == "":
-        pass
-
-    # if args.action == 'block':
-    #     if args.target:
-    #         if args.target == '-all':
+    if args.action == 'block':
+        if args.target:
+    #         if args.target == '--all':
     #             if args.sites:
     #                 command = BlockAllCommand(args.sites, blocking_manager, args.minutes)
     #             else:
     #                 print("Please specify sites to block with '-all' using the '-s' option.")
     #         else:
-    #             command = BlockSiteCommand(args.target, blocking_manager, args.minutes)
-    #     else:
-    #         print("Please specify a website to block or use '-all'.")
+            command = BlockSiteCommand(blocking_manager, args.target)
+        else:
+            print("Please specify a website to block.")
 
-    # elif args.action == 'unblock':
-    #     if args.target:
+    elif args.action == 'unblock':
+        if args.target:
     #         if args.target == '-all':
     #             if args.sites:
     #                 command = UnblockAllCommand(args.sites, blocking_manager)
     #             else:
     #                 print("Please specify sites to unblock with '-all' using the '-s' option.")
     #         else:
-    #             command = UnblockSiteCommand(args.target, blocking_manager)
-    #     else:
-    #         print("Please specify a website to unblock or use '-all'.")
+            command = UnblockSiteCommand(blocking_manager, args.target)
+        else:
+            print("Please specify a website to unblock.")
 
-    # elif args.action == 'restore':
-    #     command = RestoreHostsCommand(hosts_file, original_hosts_content)
-    #     # Optionally clear the blocking manager's state on restore
-    #     blocking_manager.indefinitely_blocked.clear()
-    #     blocking_manager.temporarily_blocked.clear()
-    #     blocking_manager._save_state()
+    elif args.action == 'list':
+        command = ListBlockedSitesCommand(blocking_manager)
 
-    # elif args.action == 'list':
-    #     command = ListBlockedSitesCommand(blocking_manager)
-
-    # if command:
-    #     command.execute()
+    if command:
+        command.execute()
